+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `IlpCredentials` is now deprecated. Please use `ILPCredentials` instead.
 
 ## 3.1.1 - 2020-05-15
+- A new method, `getPayment`, added to `XRPClient` for retrieving payment transactions by hash.
 
 This fix release contains minor updates and performance improvements.